% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_barplot.r
\name{sp_barplot}
\alias{sp_barplot}
\title{Generating bar plot}
\usage{
sp_barplot(
  data,
  variable = "Gene",
  value = "Expr",
  xvariable = "ID",
  melted = TRUE,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  level_i = c(),
  x_level = c(),
  x_type = TRUE,
  scaleY = FALSE,
  y_add = 0,
  scaleY_x = "scale_y_log10()",
  facet = "NoMeAnInGTh_I_n_G_s",
  stat = "identity",
  position = "stack",
  facet_level = NULL,
  nrow = NULL,
  ncol = NULL,
  error_bar = "ctCTct",
  base_font_size = 11,
  par = NULL,
  legend.position = "right",
  xtics = TRUE,
  xtics_angle = 0,
  ytics = FALSE,
  color = FALSE,
  color_v = c(),
  scales = "fixed",
  coordinate_flip = FALSE,
  add_text = FALSE,
  font_type = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab seperated)}

\item{variable}{Name for legend column. Default 'variable'.}

\item{value}{Name for value column. Default 'value'.}

\item{xvariable}{Name for x-axis variable.}

\item{melted}{When true, it will skip melt preprocesses. But the format must be When true, it will skip melt preprocesses. But the format must be. Default FALSE, accept TRUE.}

\item{title}{Title of picture.}

\item{x_label}{Xlab label.Default NULL.}

\item{y_label}{Ylab label.Default NULL.}

\item{x_level}{Levels for x-axis variable, suitable when x-axis is not used as a number.}

\item{x_type}{The attribute of x-axis variable. Default TRUE, means X-axis label is text. FALSE means X-axis label is number.}

\item{scaleY}{Scale y axis. Default FALSE. Accept TRUE. If the supplied number after \code{y_add} is not 0, this parameter will be automatically TRUE.}

\item{y_add}{A number to add if scale is used. Default 0. If a non-zero number is given, \code{scaleY} is TRUE}

\item{scaleY_x}{If scale is TRUE, give the following scale_y_log10(), coord_trans(y="log10"), or other legal command for ggplot2.}

\item{facet}{Wrap plots by given column. This is used to put multiple plot in one picture. Used when \code{melted} is TRUE, normally a string set should be suitable for this parameter.}

\item{stat}{The ways to show the height of bars.}

\item{position}{The ways to place multiple bars for one group if there are. Multiple bars in same place will be stacked together by default.}

\item{facet_level}{The levels of wrapping to set the order of each group.}

\item{nrow}{The number of rows one want when \code{melted} is used. Default NULL.}

\item{ncol}{The number of columns one want when \code{melted} is used. Default NULL.}

\item{error_bar}{Error-bar column. Specify the column containing error bars.}

\item{base_font_size}{Font-size. Default 11.}

\item{par}{Other legal R codes for gggplot2 will be given here.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xtics}{Display xtics. Default TRUE.}

\item{xtics_angle}{Rotation angle for x-axis labels (anti clockwise), Default 0.}

\item{ytics}{Display ytics. Default FALSE.}

\item{color}{Manually set colors for each bar. Default FALSE, meaning using ggplot2 default.}

\item{color_v}{Color for each bar.}

\item{scales}{Paramter for scales for facet. Optional, only used when \code{melted} is given. Default each inner graph use same scale (x,y range). 'free','free_x','free_y' is accepted.}

\item{coordinate_flip}{Flip cartesian coordinates so that horizontal becomes vertical, and vertical, horizontal. This is primarily useful for converting geoms and statistics which display y conditional on x, to x conditional on y.}

\item{add_text}{Add text to bar. Default FALSE.}

\item{font_type}{Specify font type. Give a path for one font type file like '/etc/fonts/Arial.ttf'
or 'HeiArial.ttc'(if in current directory), Default system default.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating bar plot
}
\examples{

bar_test_data <- data.frame(ID = letters[1:4], Gene = letters[c(8,8,9,9,10,10,11,11)], Expr = runif(16))
sp_barplot(data = bar_test_data, xvariable = "ID", value = "Expr", variable = "Gene")

## Not run:
bar_data = "bar.data"

sp_barplot(data = bar_data, xvariable = "ID", value = "Expr", variable = "Gene")
## End(Not run)

}
