% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_histogram.R
\name{sp_histogram}
\alias{sp_histogram}
\title{Generating histogram plot}
\usage{
sp_histogram(
  data,
  melted = FALSE,
  yAxis_variable,
  legend_variable,
  stat = "density",
  scale = 1,
  level_i = c(),
  facet = "haha",
  facet_order_i = c(),
  j_facet_order_i = c(),
  largest = "Inf",
  type_hist = "count/sum(count)",
  type_p = "line",
  adjust = NULL,
  pos = "identity",
  alpha = 0.4,
  fill_area = FALSE,
  line_size = 1,
  vline = FALSE,
  color_v = c(),
  j_facet = "haha",
  facet_scale = "fixed",
  facet_ncol = 1,
  xtics = TRUE,
  ytics = TRUE,
  legend.position = "right",
  xtics_angle = 90,
  xtics_value = 0,
  xtics_v = 0,
  custom_vline_coord = NULL,
  custom_vline_anno = NULL,
  x_label = NULL,
  y_label = NULL,
  title = "",
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab seperated)}

\item{melted}{When FALSE, it will skip melt preprocesses.}

\item{yAxis_variable}{Specify the column name used as value variable.}

\item{legend_variable}{Specify the column name used as legend variable.}

\item{scale}{Scale value. All values will be divided by supplied one. Default 1.}

\item{level_i}{Levels for legend variable. Accept a string like "'ctcf','h3k27ac'".}

\item{facet}{Variable name for facet. Optional, the name of one column representing a group should be given if group
information is needed. Here 'set' can be given if you want to plot set-1 and set-2 separatelt.}

\item{facet_order_i}{Order of s_facet (facet_variable given to \code{facet}). Optional even -s is given. Only specified if the facet order is what you want. Input format likes
"'facet1','facet2',"facet10"".}

\item{j_facet_order_i}{Order of J_facet (facet variable given to \code{j_facet}). Optional even -s is given. Only specified if the facet order is what you want. Input format likes
"'facet1','facet2',"facet10"".}

\item{largest}{The largest value allowed. Values other than this will be set as this value. Default Inf.}

\item{type_hist}{Plot with density or count or frequency in y-axis. Default frequency, accept density, count. When \code{type_p} is both, frequency will be given here.}

\item{type_p}{Plot frequency or density or hist or both frequency and histogram. Default line means frequency accept density_line, hist or both.}

\item{adjust}{The value for adjust (like the width of each bin) in geom_density, and binwidth for geom_histogram.}

\item{pos}{Position paramter for hist bars. Default identity, accept dodge.}

\item{alpha}{Alpha value for transparent. Default 0.4, accept a number form 0 to 1,  the smaller, the more transparent.}

\item{fill_area}{Fill the area if TRUE. Default FALSE.}

\item{line_size}{line size. Default 1. Accept a number.}

\item{vline}{Add mean value as vline. Default FALSE,  accept TRUE.}

\item{color_v}{Color for each line. Str in given
format must be supplied, ususlly the number of colors should
be equal to the number of lines.
"'red','pink','blue','cyan','green','yellow'" or
"rgb(255/255,0/255,0/255),rgb(255/255,0/255,255/255),rgb(0/255,0/255,255/255),
rgb(0/255,255/255,255/255),rgb(0/255,255/255,0/255),rgb(255/255,255/255,0/255)"}

\item{j_facet}{Another variable for facet. Optional, same meaning as \code{facet} but different columns, like 'variable'. When this is given,
facet_grid would be used in format like facet_grid(s_facet ~ j_facet) pay attention to the order.}

\item{facet_scale}{Parameter for scales for facet. Necessary if only \code{facet} is given. Default each inner graph use same scale (x, y range). "free", 'free_x', 'free_y' is accepted.}

\item{facet_ncol}{Number of columns in one row when do faceting. Necessary if only \code{facet} is given}

\item{xtics}{Show the X axis.}

\item{ytics}{Show the Y axis.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xtics_angle}{Rotation angle for a-axis.}

\item{xtics_value}{Manually set the value of xtics when \code{xtics_v} is specified. Default the content of \code{xtics_v} when \code{xtics_v} is specified,
accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector to set the position of xtics.}

\item{xtics_v}{Manually set xtics. Default FALSE, accept a series of
numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector.}

\item{custom_vline_coord}{Add vline by given point. Default FALSE, accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector.}

\item{custom_vline_anno}{Add labels to vline. Default FALSE, accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector.}

\item{x_label}{Xlab label.}

\item{y_label}{Ylab label.}

\item{title}{Title of picture.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating histogram plot
}
\examples{

histogram_test_data <- data.frame(Type = letters[1:2], Value = runif(80))
sp_histogram(data = histogram_test_data, yAxis_variable = "Value", legend_variable = "Type",
 type_p = "both", type_hist = "count/sum(count)")

## Not run:
input <- "histogram.data"
sp_histogram(data=input, yAxis_variable = "Value", legend_variable = "Type",
 type_p = "both", type_hist = "count/sum(count)")
## End(Not run)
}
