% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lines.r
\name{sp_lines}
\alias{sp_lines}
\title{Generating lines plot}
\usage{
sp_lines(
  data,
  melted = TRUE,
  x_type = TRUE,
  xvariable = "Pos",
  y_add = 0,
  yvariable = "value",
  scaleY_x = "scale_y_log10()",
  level = NULL,
  variable = "Variable",
  color_variable = "Variable",
  color_order = NULL,
  x_level = NULL,
  x_label = NULL,
  y_label = NULL,
  title = "",
  smooth = FALSE,
  smooth_method = "auto",
  line_size = NULL,
  scaleY = FALSE,
  color = FALSE,
  color_v = c(),
  xtics = TRUE,
  xtics_angle = 0,
  ytics = "TRUE",
  legend.position = "right",
  xtics_pos = 0,
  xtics_value = 0,
  uwid = 0,
  vhig = 0,
  xintercept = NULL,
  custom_vline_anno = NULL,
  yintercept = NULL,
  custom_hline_anno = NULL,
  facet = NULL,
  ...
)
}
\arguments{
<<<<<<< HEAD
\item{data}{Data file (with header line, the first column would be be treated as rownames for normal matrix. No rownames for melted format. Columns are tab seperated).}
=======
\item{data}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab seperated)}
>>>>>>> 0699e6abe6ebc2c923820a14f12ffbeb64cf70a6

\item{melted}{When true, it will skip melt preprocesses. But the format must be When true, it will skip melt preprocesses. But the format must be. Default FALSE, accept TRUE.}

\item{x_type}{Are x-axis variables numbers. Default TRUE, meaning X-axis label is text. FALSE means X-axis label is numerical.}

\item{xvariable}{Name for x-axis variable,For normal matrix, default the first column will be used,program will assign an value 'xvariable' to represent it.}

\item{y_add}{A number to add if scale is used.A number to add if scale is used.}

\item{yvariable}{Name for y-axis variable.Default value.}

\item{scaleY_x}{If scaleY is TRUE, give the following \code{scale_y_log10()} (default), \code{coord_trans(y="log10")},  or other legal command for ggplot2 or simply \code{log2}.}

\item{level}{Set orders of legend variable. Default column order for normal matrix, accept a string like c('CTCF','H3K27ac','Enhancer') to set your own order.}

\item{variable}{Name for legend variable. Default variable, this should only be set when \code{melted} is TRUE.}

\item{color_variable}{Name for color variable. Normally used when wanting to color lines by groups. Default same as \code{variable}, this should only be set when \code{melted} is TRUE.}

\item{color_order}{Set orders of color variable. Only used when \code{color_variable} is not the same as \code{variable}.}

\item{x_level}{Levels for x-axis variable, suitable when x-axis is not treated as numerical. Default the order of first column for normal matrix. This will only be considered when \code{x_type} is TRUE.}

\item{x_label}{Xlab label.Default empty.}

\item{y_label}{Ylab label.Default empty.}

\item{smooth}{Smooth lines or not. Default FALSE means no smooth. Accept TRUE to smooth lines.}

\item{smooth_method}{The smooth method you want to use. Smoothing method (function) to use, eg. lm, glm, gam, loess,rlm. observations < 1000 default is 'loess', observations >= 1000 defaults to 'gam'.}

\item{line_size}{line size. Default 1. Accept a number.}

\item{scaleY}{Scale y axis. Default null. Accept TRUE. This function is depleted.}

\item{color}{Manually set colors for each line. Default FALSE, meaning using ggplot2 default.}

\item{color_v}{Color for each line.}

\item{xtics}{Default TRUE. Display xtics. FALSE don't display.}

\item{xtics_angle}{Rotation angle for x-axis labels (anti clockwise), Default 0.}

\item{ytics}{Default TRUE. Display ytics. FALSE don't display.}

\item{xtics_pos}{Manually set the position of xtics. Default FALSE, accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector to set the position of xtics.}

\item{xtics_value}{Manually set the value of xtics when \code{xtics_pos} is specified.
Default the content of \code{xtics_pos} when \code{xtics_pos} is specified, accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector to set the position of xtics.}

\item{facet}{The formula for facets. Default no facets,
\code{+facet_grid(level ~ .)} means divide by levels of 'level' vertically.
\code{+facet_grid(. ~ level)} means divide by levels of 'level' horizontally.
\code{+facet_grid(lev1 ~ lev2)} means divide by lev1 vertically and lev2 horizontally.
\code{+facet_wrap(~level, ncol=2)} means wrap horizontally with 2 columns.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating lines plot
}
\examples{

res_output <- data.frame(Pos=1:10,value =runif(20))
value=0.5
res_output$Variable <- ifelse(res_output$value<=value,"groupA", "groupB")
sp_lines(data=res_output, color_variable = "Variable", variable="Variable")


## Not run:
lines_data = "lines.data"

sp_lines(data=lines_data, xvariable = "Pos", yvariable = "value", color_variable = "Variable", variable="Variable")
## End(Not run)

}
