% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pca.R
\name{sp_pca}
\alias{sp_pca}
\title{Generating pca plot}
\usage{
sp_pca(
  data,
  grp_file = NULL,
  title = "",
  scale = TRUE,
  color = "c_t_c_t0304",
  color_v = c(),
  log = "nolog",
  facet = NULL,
  size = NULL,
  shape = "c_t_c_t0304",
  color_order = c(),
  top_n = 5000,
  shape_order = c(),
  dimensions = 2,
  alpha = 1,
  mid = ".pca",
  label = FALSE,
  log_add = 1,
  label_font_size = NULL,
  minimum_mad = 0.5,
  debug = FALSE,
  saveplot = NULL,
  filename = "data",
  ...
)
}
\arguments{
\item{data}{Data file. With header line, the first column is the rowname, tab seperated. Each row represents variable (normally genes), each column represents samples.}

\item{grp_file}{Sample group file with first column as sample names, other columns as sample attributes. Below, color, size, shape variable should be existed in this file.}

\item{title}{Title of picture. Default empty title}

\item{scale}{Scale data for PCA analysis. In defaut, prcomp will centralized data by minus mean value and
normalize data by column standard deviation dividing.
Often, we would normalize data. Only when we care about the real number changes other than the trends,
we do not need scale. When this happens, we expect the changin ranges of data is small for example log-transformed data.}

\item{color}{The variable for point color. Optional, such as color.}

\item{color_v}{Manually specified colors. Default system default. Accept string in format like <'"green", "red"'> (both types of quotes needed).}

\item{log}{Log-transform data before principle component analysis.}

\item{facet}{The formula for facets. Default no facets,
facet_grid(level ~ .) means divide by levels of 'level' vertcally.
facet_grid(. ~ level) means divide by levels of 'level' horizontally.
facet_grid(lev1 ~ lev2) means divide by lev1 vertically and lev2 horizontally.
facet_wrap(~level, ncol=2) means wrap horizontally with 2 columns.
Example: facet_wrap(~Size,ncol=6,scale='free')}

\item{size}{The variable for point size. Optional, such as a number or a variable like count, normally should be number column.}

\item{shape}{The variable for point shape. Optional, such as shape.}

\item{color_order}{The order for color variable. Default alphabetical order, accept a string like "'K562','hESC','GM12878','HUVEC','NHEK','IMR90','HMEC'".}

\item{top_n}{Use top n most changed variables for PCA computation. Default 5000. Giving 0 to use all variables.}

\item{shape_order}{The order for shape variable. Default alphabetical order, accept a string like "'K562','hESC','GM12878','HUVEC','NHEK','IMR90','HMEC'".}

\item{dimensions}{Dimensions to plot. Default 2.    Accept 3 (both color and shape variable needed and should be same variable).}

\item{alpha}{Transparency value for points. Optional, such as a number or a variable indicating one data column, normally should be number column.}

\item{label}{Label points (using geom_text_repel). Default no-label (FALSE), accept TRUE.}

\item{log_add}{Add a value before log-transform to avoid log(0). Default 1.}

\item{label_font_size}{Label font size. Default system default. Accept numbers less than 5 to shrink fonts.}

\item{minimum_mad}{Minimum mad to keep. Larger mad, larger variance. Default 0.5.}

\item{...}{}
}
\value{
pdf and xls files.
}
\description{
Generating pca plot
}
\examples{
pca_test_data <- matrix(runif(3000,0,100000),ncol=6)
colnames(pca_test_data) <- c(paste("wt",1:3,sep = ""),paste("ko",1:3,sep = ""))
rownames(pca_test_data) <- c(ID = paste0("ENSG",c(1:500)))
pca_data <- as.data.frame(pca_test_data)
sp_pca(data = pca_data, grp_file = NULL)


## Not run:
data = "pca.data"
group_data = "pca_group.data"
sp_pca(data = data, grp_file = group_data, color="Conditions", size = "Diameters", shape = "Batch", label = TRUE)
## End(Not run)


}
