% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pheatmap.R
\name{sp_pheatmap}
\alias{sp_pheatmap}
\title{Generating pheatmap plot}
\usage{
sp_pheatmap(
  data = "pheatmap.txt",
  deal_duplicateRname = 0,
  logv = NULL,
  scale = "none",
  gradient = 1,
  mid_value = "Inf",
  digits = FALSE,
  givenSepartor = NULL,
  annotation_row = NULL,
  annotation_col = NULL,
  clustering_method = "complete",
  clustering_distance_rows = "correlation",
  clustering_distance_cols = "correlation",
  maximum = Inf,
  minimum = -Inf,
  xtics_angle = 270,
  color_type = "function",
  color_vector = "colorRampPalette(rev(brewer.pal(n=7, name=\\"RdYlBu\\")))(100)",
  bias = 1,
  fontsize = 14,
  annotation_colors = "NA",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  cor_plot = "None",
  kclu = NA,
  ytics = TRUE,
  xtics = TRUE,
  title = "",
  ...
)
}
\arguments{
\item{data}{Data file (with header line, the first column is the rowname, tab seperated. Colnames must be unique unless you know what you are doing.)}

\item{deal_duplicateRname}{Specify the way to deal with duplicate row names.
Default 0: representing duplicated row names are not allowed.
Accept  1: representing make duplicated row names unique by adding <.1>, <.2> for the second, third appearances.}

\item{logv}{}

\item{scale}{}

\item{gradient}{}

\item{mid_value}{}

\item{digits}{}

\item{givenSepartor}{}

\item{annotation_row}{}

\item{annotation_col}{}

\item{clustering_method}{Clustering method, Default "complete". Accept "ward.D", "ward.D2","single", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or "centroid" (=UPGMC)}

\item{clustering_distance_rows}{}

\item{clustering_distance_cols}{}

\item{maximum}{}

\item{minimum}{}

\item{xtics_angle}{Rotation angle for x-axis value. Default 90.}

\item{color_type}{Color type, a vetcor which will be transferred as described in \code{color_vector} or a raw vector or    a function (default).}

\item{color_vector}{Color vector. Default pheatmap_default. Aceept a vector containing multiple colors such as 'c("white", "blue")' will be transferred
to colorRampPalette(c("white", "blue"), bias=${bias})(30)  or an R function colorRampPalette(rev(brewer.pal(n=7, name="RdYlBu")))(100)  generating a list of colors.}

\item{bias}{}

\item{fontsize}{}

\item{annotation_colors}{}

\item{cluster_rows}{}

\item{cluster_cols}{}

\item{cor_plot}{Do correlation plot. Default "None", accept "row" or "col" for row correlation or column correlation.}

\item{kclu}{}

\item{ytics}{Display ytics.}

\item{xtics}{Display xtics.}

\item{title}{Title of picture. Default empty title}

\item{...}{}
}
\value{

}
\description{
Generating pheatmap plot
}
