% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_upsetView.R
\name{sp_upsetview}
\alias{sp_upsetview}
\title{Generating upsetView plot}
\usage{
sp_upsetview(
  data,
  vennFormat = 0,
  pointsize = 8,
  keep_empty = TRUE,
  saveplot = "1.pdf",
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{Data file. Receive long and wide table forms.}

\item{vennFormat}{Venn diagram format without header line. Default 0 represents normal data. Accept 1,2.
0: represents wide data listed above.
1: represents venn diagram format without header line.
2: represents venn diagram format with header line.}

\item{pointsize}{Point size. Default 8.}

\item{keep_empty}{Keep empty intersections. Default TRUE. Accept FALSE to remove empty intersections.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{}
}
\value{
A pdf file.
}
\description{
Generating upsetView plot
}
\examples{

upsetview_data <- data.frame(elements=c("1","2","2","2","3"), sets=c("A","A","B","C","C"))
sp_upsetview(data = upsetview_data, vennFormat=2, saveplot = "upsetView_long.pdf")


## Not run:
upsetview_data = "upsetview.data"
sp_upsetview(data = upsetview_data, saveplot = "upsetView_wide.pdf")
## End(Not run)

}
