% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_vennDiagram3.R
\name{sp_vennDiagram3}
\alias{sp_vennDiagram3}
\title{Generating venDiagram plot}
\usage{
sp_vennDiagram3(
  data,
  header = TRUE,
  item_variable = NULL,
  set_variable = NULL,
  select_set_to_show = NULL,
  doWeights = FALSE,
  type = "ellipses",
  SetLabels = TRUE,
  Faces = TRUE,
  Sets = TRUE,
  FaceText = "weight",
  saveplot = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{Data file (the first column is the name of genes or other things one want to compare, the second column is set name of each items, tab separated).}

\item{header}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}

\item{item_variable}{Specify the column containing all items (one of column names of data).}

\item{set_variable}{Specify the column containing set names (one of column names of data).}

\item{doWeights}{Whether to use weights to represent data sets}

\item{type}{Represent data sets in different shapes.}

\item{SetLabels}{Whether to plot the names of the Sets. Default TRUE.}

\item{Faces}{If Faces = TRUE, the sets will be filled with colors.}

\item{Sets}{If Sets = TRUE, the boundaries of the Sets are shown.}

\item{FaceText}{FaceText is a character vector which may contain any of c('weight','signature','sets','elements').}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Other parameters given to \code{base_plot_save}}
}
\value{
A pdf file.
}
\description{
Generating venDiagram plot
}
