% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_cytoscape_each_module}
\alias{WGCNA_cytoscape_each_module}
\title{Output nodes and edges data for each module only. (deprecated)}
\usage{
WGCNA_cytoscape_each_module(edgeData, nodeData, module, prefix)
}
\arguments{
\item{edgeData}{generated by \code{WGCNA_cytoscape}}

\item{nodeData}{generated by \code{WGCNA_cytoscape}}

\item{module}{A module color name}

\item{prefix}{prefix for output files.}
}
\description{
Output nodes and edges data for each module only. (deprecated)
}
\examples{

None

}
