% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_dataFilter}
\alias{WGCNA_dataFilter}
\title{Filter data for WGCNA input to increase computing efficiency without loosing too
many information.}
\usage{
WGCNA_dataFilter(datExpr, ...)
}
\arguments{
\item{datExpr}{Normal gene expression matrix (gene x sample).}

\item{...}{Parameters given to \code{dataFilter}.}
}
\value{
A dataframe (samples x genes).
}
\description{
Filter data for WGCNA input to increase computing efficiency without loosing too
many information.
}
\examples{

exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)

}
