% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_boxplot.R
\name{sp_boxplot}
\alias{sp_boxplot}
\title{Generating box plot}
\usage{
sp_boxplot(
  data,
  melted = FALSE,
  xvariable = "",
  value = "",
  variable = "",
  statistics = FALSE,
  xtics_angle = 0,
  level_i = c(),
  legend_cut = NULL,
  x_level = c(),
  x_cut = NULL,
  scaleY = FALSE,
  y_add = 0,
  scaleY_x = "scale_y_log10()",
  notch = FALSE,
  par = NULL,
  outlier = FALSE,
  out_scale = 1.05,
  legend.position = "right",
  color_v = NULL,
  violin = FALSE,
  violin_nb = FALSE,
  scale_violin = "width",
  ID_var = c(),
  jitter = FALSE,
  jitter_bp = FALSE,
  colormodel = "srgb",
  coordinate_flip = FALSE,
  facet = "NoMeAnInGTh_I_n_G_s",
  facet_level = NULL,
  x_label = NULL,
  y_label = NULL,
  title = NULL,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  gene = NULL,
  sampleGroup = NULL,
  ...
)
}
\arguments{
\item{data}{Data file (with header line, the first row is the colname, tab seperated. Multiple formats are allowed and described above)}

\item{melted}{When FALSE, it will skip preprocess. Default FALSE.}

\item{xvariable}{The column represents the digital values.}

\item{value}{The column represents the digital values. This parameter can only be set when \code{melted} is FALSE.}

\item{variable}{The column represents the variable information, meaning legend_variable.}

\item{xtics_angle}{Rotation angle for x-axis value. Default 0.}

\item{level_i}{Levels for legend variable. Default data order,accept input like c('TP16','TP22','TP23') for variable column.}

\item{legend_cut}{Self-define intervals for legend variable when legend is continuous numbers.}

\item{x_level}{Levels for x-axis variable. Default data order, accept input like c('g','a','j','x','s','c','o','u') for Set column.}

\item{x_cut}{Self-define intervals for x-axis variable.}

\item{scaleY}{Scale y axis. Default FALSE.  Accept TRUE. Also if the supplied number after \code{y_add} is not 0, this parameter will be set to TRUE}

\item{y_add}{A number to add if scale is used. Default 0.}

\item{scaleY_x}{If scale is TRUE, give the following scale_y_log10() (default), coord_trans(y="log10"), scale_y_continuous(trans=log2_trans()), coord_trans(y="log2"),
or other legal command for ggplot2).}

\item{notch}{Using notch (sand clock shape) or not. Default FALSE.}

\item{outlier}{Exclude outliers. Exclude outliers or not, default FALSE means not.}

\item{out_scale}{The scales for you want to zoom in to exclude outliers. Default 1.05. No recommend to change unless you know what you are doing.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{color_v}{Manually set colors for each box. Default NULL, meaning using ggplot2 default.}

\item{violin}{Do violin plot instead of boxplot.}

\item{violin_nb}{Do violin plot without inner boxplot.}

\item{scale_violin}{The value given to scale for violin plot.
if "area", all violins have the same area (before trimming the tails).
If "count", areas are scaled proportionally to the number of observations.
If "width", all violins have the same maximum width. 'equal' is also accepted.}

\item{ID_var}{Other columns you want to treat as ID variable columns except the one given to \code{xvariable}.}

\item{jitter}{Do jitter plot instead of boxplot.}

\item{jitter_bp}{Do jitter plot overlay with violinplot or boxplot or both.}

\item{coordinate_flip}{Rotate the plot from vertical to horizontal. Usefull for plots with many values or very long labels at X-axis}

\item{facet}{Wrap plots by given column. This is used to put multiple plot in one picture. Used when \code{melted} is FALSE, normally a string set should be suitable for this parameter.}

\item{facet_level}{The levels of wrapping to set the order of each group.}

\item{x_label}{Xlab label.Default NULL.}

\item{y_label}{Ylab label.Default NULL.}

\item{title}{Title of picture.}

\item{nrow}{The number of rows one want when \code{melted} is used. Default NULL.}

\item{ncol}{The number of columns one want when \code{melted} is used. Default NULL.}

\item{scales}{Paramter for scales for facet. Optional, only used when \code{x_level} is given. Default each
inner graph use same scale (x,y range).
'free' (variable x, y ranges for each sub-plot), 'free_x' (variable x ranges for each sub-plot), 'free_y'  is accepted.}

\item{gene}{Giving one gene ID to do boxplot specifically for this gene. Default FALSE, accept a string.}

\item{sampleGroup}{Giving a sampleGroup file with format specified above to tell the group information for each sample.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating box plot
}
\examples{

box_test_data <- data.frame(ID = letters[1:4], Gene = letters[c(8,8,9,9,10,10,11,11)], Expr = runif(16))
sp_boxplot(data = box_test_data, xvariable = "ID", value = "Expr", variable ="Gene")

## Not run:
box_data = "box.data"

sp_boxplot(data = box_data, xvariable = "Gene", value = "Expr", variable="Group")
## End(Not run)

}
