% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_histogram.R
\name{sp_histogram}
\alias{sp_histogram}
\title{Generating histogram plot}
\usage{
sp_histogram(
  data,
  melted = FALSE,
  yvariable = NULL,
  color_variable = NULL,
  color_variable_order = NULL,
  value_scale = 1,
  facet_variable = NULL,
  facet_variable_order = NULL,
  maximum_allowed = "Inf",
  yaxis_statistics = "count/sum(count)",
  plot_type = "line",
  adjust = NULL,
  hist_bar_position = "identity",
  alpha = 0.4,
  yaxis_scale_mode = NULL,
  y_add = 0,
  fill_area = FALSE,
  line_size = 1,
  add_mean_value_vline = FALSE,
  manual_color_vector = c(),
  facet_scale = "fixed",
  facet_ncol = 1,
  facet_nrow = 1,
  xtics = TRUE,
  ytics = TRUE,
  legend.position = "right",
  xtics_angle = 90,
  manual_xtics_value = 0,
  manual_xtics_pos = 0,
  custom_vline_x_position = NULL,
  custom_vline_anno = NULL,
  x_label = NULL,
  y_label = NULL,
  title = "",
  font_path = NULL,
  debug = F,
  extra_ggplot2_cmd = NULL,
  file_name = NULL,
  base_font_size = 11,
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first
column will not be treated as the row-names, tab separated)}

\item{melted}{\code{TRUE} for dealing with long format matrix, the program will skip
melt preprocess. If input is wide format matrix, this parameter should be set to \code{FALSE}.}

\item{yvariable}{Specify the column name used as Y-axis value (one of column names,
should be specified when inputting long format matrix).}

\item{color_variable}{Specify the column name used as color variable.}

\item{color_variable_order}{Levels for color variable to set their appearing order.
Accept a vector like \code{c('ctcf','h3k27ac')}
(normally whole or a subset of unique values of one column with order specified).}

\item{value_scale}{Scale value. All values will be divided by supplied one. Default 1.}

\item{facet_variable}{Wrap plots by given column (one of column names should be specified).
This is used to put multiple plot in one picture.}

\item{facet_variable_order}{Order of facets (to set the appearance order of each subplot).
Accept a vector like \code{c('ctcf','h3k27ac')}
(normally whole or a subset of unique values of one column with order specified).}

\item{maximum_allowed}{The largest value allowed. Values other than this will be set
as this value. Default Inf.}

\item{yaxis_statistics}{Plot with density or count or frequency in y-axis.
Default \code{frequency}, accept \code{density}, \code{count}.
When \code{plot_type} is both, `frequency`` will be given here.}

\item{plot_type}{Plot frequency or density or hist or both frequency and histogram.
Default line means frequency accept \code{density_line}, \code{hist} or \code{both}.}

\item{adjust}{The value for adjust (like the width of each bin) in
\code{geom_density}, and \code{binwidth} for \code{geom_histogram}.}

\item{hist_bar_position}{Position parameter for hist bars. Default \code{identity}, accept \code{dodge}.}

\item{yaxis_scale_mode}{Give the following \code{scale_y_log10()},
\code{coord_trans(y="log10")}, or other legal command for ggplot2 or
simply \code{log2} to set the scale way.}

\item{y_add}{A number to add if log scale is used.
Default 0 meaning the minimum non-zero value would be used.}

\item{fill_area}{Fill the area if TRUE. Default FALSE.}

\item{line_size}{line size. Default 1. Accept a number.}

\item{add_mean_value_vline}{Add mean value as vline. Default FALSE,  accept TRUE.}

\item{facet_ncol}{The number of columns one want when \code{facet} is used. Default NULL.}

\item{facet_nrow}{The number of rows one want when \code{facet} is used. Default NULL.}

\item{xtics}{Show the X axis.}

\item{ytics}{Show the Y axis.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xtics_angle}{Rotation angle for a-axis. Default 0.}

\item{manual_xtics_value}{Manually set the values of xtics when \code{manual_xtics_position} is specified.
Default the content of \code{manual_xtics_position} when \code{manual_xtics_position} is specified,
accept a series of numbers in following format "c(1,2,3,4,5)" or other R code that can
generate a vector to set the value of manual xtics.}

\item{manual_xtics_pos}{Manually set the positions of xtics. Default FALSE, accept a series of
numbers in following format "c(1,2,3,4,5)" or other R code that can generate a vector
to set the position of manual xtics.}

\item{custom_vline_x_position}{A vector of coordinates for vertical lines.}

\item{custom_vline_anno}{Annotation text for each vertical line.}

\item{x_label}{Xlab label.}

\item{y_label}{Ylab label.}

\item{title}{Title of picture.}

\item{font_path}{Specify font type. Give a path for one font type file
like '/etc/fonts/Arial.ttf'
or 'HeiArial.ttc'(if in current directory), Default system default.}

\item{extra_ggplot2_cmd}{Extra ggplot2 commands (currently unsupported)}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating histogram plot
}
\examples{

histogram_test_data <- data.frame(Type = letters[1:2], Value = runif(80))
sp_histogram(data = histogram_test_data, yvariable = "Value", color_variable = "Type",
 plot_type = "both", yaxis_statistics = "count/sum(count)")

## Not run:
input <- "histogram.data"
sp_histogram(data=input, yvariable = "Value", color_variable = "Type",
 plot_type = "both", yaxis_statistics = "count/sum(count)")
## End(Not run)
}
